+++ tools/quadrados/src/tools.hpp
@@ -17,10 +17,12 @@ int runHelp(int argc, char** argv, GlobalArgs& ga);
 int runEmbed(int argc, char** argv, GlobalArgs& ga);
 int runConvert(int argc, char** argv, GlobalArgs& ga);
 int runInit(int argc, char** argv, GlobalArgs& ga);
+int runCreate(int argc, char **argv, GlobalArgs& ga);
 
 static const Tool Tools[] = {
     {"help", runHelp},
     {"embed", runEmbed},
     {"convert", runConvert},
     {"init", runInit},
+    {"create", runCreate},
 };